fastsrv/src/requirements.txt
pyyaml==3.12
Jinja2==2.8
Django==1.10.3
